requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==23.0.0
whitenoise>=6.6.0
psycopg[binary]>=3.2.0
